requirements.txt
requests
python-bitcoinlib
ecdsa
base58
pycryptodome